+++ packages/cookie-banner/src/components/register.ts
@@ -30,7 +30,6 @@ import { ProboCategoryToggle } from "./category-toggle";
 import { ProboCookieBannerRoot } from "./cookie-banner-root";
 import { ProboCookie, ProboCookieList } from "./cookie-list";
 import { ProboPreferencePanel, ProboSaveButton } from "./preference-panel";
-import { ProboSettingsButton } from "./settings-button";
 import { ProboSettingsLink } from "./settings-link";
 
 const elements: [string, CustomElementConstructor][] = [
@@ -46,7 +45,6 @@ const elements: [string, CustomElementConstructor][] = [
   ["probo-cookie-list", ProboCookieList],
   ["probo-cookie", ProboCookie],
   ["probo-save-button", ProboSaveButton],
-  ["probo-settings-button", ProboSettingsButton],
   ["probo-settings-link", ProboSettingsLink],
 ];